Fix refactor async operation

+++ novawallet/Common/Services/ChainRegistry/RuntimeProviderPool/RuntimeSyncService.swift
@@ -1,5 +1,5 @@
 import Foundation
-import RobinHood
+import Operation_iOS
 import SubstrateSdk
 
 protocol RuntimeSyncServiceProtocol {
@@ -63,7 +63,7 @@ final class RuntimeSyncService {
         operationQueue: OperationQueue,
         retryStrategy: ReconnectionStrategyProtocol = ExponentialReconnection(),
         dataHasher: StorageHasher = .twox256,
-        rpcTimeout: Int = Int.max,
+        rpcTimeout: Int = Int(UInt16.max),
         logger: LoggerProtocol? = nil
     ) {
         self.repository = repository
